Add optimize_model_load setting to speedup loading of model objects, off by default
This setting is off by default as it isn't completely compatible. But
it can speed up model loading by 30-50% for simple tables.
Basically, this is just like the standard way of returning a hash,
except that instead of yielding the hash, a model object is
allocated, the hash is assigned as the values, and the model object
is yielded.