Get microsecond accuracy when using datetime_class = DateTime with 1.8-1.9.2 stdlib date library via Rational
Unfortunately, with the 1.8-1.9.2 stdlib date class, you can't add a
float to a DateTime instance and get microsecond accuracy, even if the
float is very small.  For a current DateTime instance, accuracy appears
to be +/- 33 usecs.

To work around this issue, check if the 1.8-1.9.2 stdlib date
implementation is being used via the @ajd instance variable.  If so,
use a slower but more accurate Rational implementation.

While here, register spg_SQLTime and spg_Postgres with the garbage
collector to avoid segfaults if the constants are manually undefined.
Also, rename a misleading macro to reflect it uses microseconds instead
of milliseconds.