Add PostgreSQL array parser
This adds a C-based PostgreSQL array parser.  The original
implementation is from the pg_array_parser library, but I've
heavily modified it.  This C-based parser is 5-500 times faster
than the pure ruby parser that Sequel uses by default.  5 times
faster for an empty array, and 500 times faster for an array
with a single 10MB string.

Because the pg_array extension can be loaded before or after
sequel_pg, handle the case where it is loaded after by switching
the Creator class to use the C based parser instead of the pure
ruby parser.