Replace PQsetRowProcessor streaming with PQsetSingleRowMode streaming introduced in PostgreSQL 9.2beta3
PostgreSQL decided that PQsetRowProcessor was a bad API, and replaced
it with PQsetSingleRowMode in 9.2beta3.  I agree with the choice, as
the new API is simpler and fairly easy to use.

Currently, this requires a patch to ruby-pg to recognize
PGRES_SINGLE_TUPLE as an OK result status.  That should hopefully
be committed soon.  This also requires the master branch of Sequel,
as it depends on some recent refactoring in the Sequel postgres
adapter.