Make streaming support not swallow errors when rows are not retrieved, bump version to 1.6.3
If the user doesn't retrieve all of the rows when streaming, sequel_pg
is supposed to flush the connection by manually retrieving the other
rows.  Before, it wasn't checking the result status of each result,
causing it to miss errors.

I noticed this after adding an integration test for deferred foreign
key support to Sequel.