Work with ruby pg 0.18+ (currently a prerelease), bump version to 1.6.11
ruby-pg 0.18 changes the internals so that the data objects that
were previously pointers to PGconn and PGresult are now structs
that contain pointers to PGconn or PGresult.

If ruby-pg 0.18+ is detected (currently using
defined?(PG::TypeMapAllStrings)), then this changes the code to
use ruby-pg C functions that return the underlying pointers.

The only other change is the error message when too many columns
are used now includes the actual number of columns, instead of
just the maxinum number allowed.