prevent premature GC of string buffer in parse_pg_array
To ensure objects are properly marked, RB_GC_GUARD must be
placed after the last use of pointers accessed from the
object (in this case, the result of RSTRING_PTR).

So, instead of passing the result of RSTRING_PTR to read_array,
pass the String object itself and move the RB_GC_GUARD to
the end of the read_array function.

Warning: I've only compile-tested, not actually run since there's no
tests bundled.

Sidenote: I'm fairly sure the rest of the existing RB_GC_GUARD
calls are unnecessary.