Improve performance of timestamp parsing by borrowing and optimizing ruby-pg's parser
This uses ruby-pg's timestamp parser as a base, but makes some
changes.  It includes length checks before all pointer derefencing
while parsing the timestamp string.  It skips the isdigit checking
for year/month/day/hour/minute/second, only checking that the
delimiter values are correct.  It only checks for 6 digits of
fractional second precision, since that is the most that
PostgreSQL uses.  It checks for up to 6 digit years, since
PostgreSQL supports years in timestamps up to 294276.  It also
supports BC dates, since PostgreSQL supports years in timestamps
back to 4713 BC.

Instead of checking for the application timezone, database timezone,
and datetime class on every parse, it checks for them only once and
stores the information in a bitmap which it passes to the function.
To handle the timestamp array parsing case, it stores the bitmap
in the VALUE used for conversion procs, which is unused for types
that are natively handled (and the timestamp arrays are two of those
types).

This includes a change so that conversion procs are not checked
for natively handled array types, since they would be unused.

This provides the most speedup on ruby 2.3+, as
ruby_time_timespec_new is used there.  It still should be faster
on ruby <2.3 as the handwritten parser is faster than the sscanf
implementation used previously.

Previously, the code supported datetime_classes other than Time
and DateTime, using the DateTime API if the datetime_class was
not Time.  Now, if the datetime_class is not Time, sequel_pg is
hard-coded to use DateTime.