Improve performance of time parsing
This takes the same basic approach as the timestamp parsing. Instead
of getting the current time and taking the year, month, and day
from that for every parse, it gets the information up front once
per result set, and passes it via the conversion proc VALUE, storing
the year, month, and day bitshifted into a single int.

Rename the tz_info variable to timestamp_info while here.