Fix timezone handling
The previous code was buggy, but it was hidden because the
bitmap handling forced the slow pure-ruby path if timezones
were specified.

Instead of trying to cram the custom timezone by setting
both DB and LOCAL in the bitmap, just use unique values.
While it shouldn't matter, use | instead of + to modify
the bitmap.

Switch to using timegm if DB is UTC, and set offset_seconds
to INT_MAX-1 if APP is UTC, which gets us the behavior we
want without rb_funcall.