Simplify inet/cidr array type conversion
In spg__col_value, allow overridding of inet/cidr type conversion
using conversion procs.  Use goto to jump to the default case if
a conversion proc is set for the array type.

In spg__array_col_value, use convertor argument, and fallthrough
to keep things simpler and avoid encoding issues.  The fallthrough
will probably not happen in the typical case due to changes in
spg__col_value, but if parse_pg_array is called directly, it
probably could be reached.