Move PQclear to the result stream loop in pg, to avoid double free
sequel_pg-1.16.0 introduced PQclear to fix a memory leek, but now a double free can happen, if there's an interrupt between PQclear and re-assiging a new PQresult.
With this change the yielder can decide, if the result is cleared or not in the stream loop.
In this case pgresult_clear() is called instead of PQclear, which avoids a double free.

The necessary change of pgresult_stream_any() is introduced in pg-1.4.4, so that this version is now required.

Fixes https://github.com/ged/ruby-pg/issues/481