Optimize Dataset#all and #with_sql_all
This allocates an array with the expected size up front, and avoids
a C -> Ruby call per row.

Instead of checking whether Sequel::Model::ClassMethods is already
defined during load, check it inside optimize_model_load. This
allows for optimization in the case where sequel/core is required,
then the database is setup, and then sequel is required.

However, this approach also results in unnecessary slowdown of
Dataset#each when models are not used. An alternative approach will
be used in a future commit to avoid this.